backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
pydantic>=2.7.0
pyjwt>=2.8.0
python-multipart>=0.0.9

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
factory-boy>=3.3.0
faker>=19.0.0
